requirements.txt
alembic==0.9.9
appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
bcrypt==3.1.4
cffi==1.11.5
click==6.7
cryptography==2.2.2
decorator==4.3.0
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
idna==2.6
ipython==6.4.0
ipython-genutils==0.2.0
itsdangerous==0.24
jedi==0.12.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.0
netmiko==2.1.1
paramiko==2.4.1
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.1.0
prompt-toolkit==1.0.15
ptyprocess==0.5.2
pyasn1==0.4.3
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.2.1
pyserial==3.4
python-dateutil==2.7.3
python-editor==1.0.3
PyYAML==3.12
scp==0.11.0
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.8
textfsm==0.4.1
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.2.1
